backend/requirements.txt
flask
flask-cors
tensorflow
numpy
opencv-python
pillow
